// packages/server/src/scraper/enum/VisaRequirement.enum.ts
export enum VisaRequirementType {
  VISA_FREE = "VISA_FREE",
  VISA_ON_ARRIVAL = "VISA_ON_ARRIVAL",
  ETA = "ETA",
  EVISA = "EVISA",
  VISA_REQUIRED = "VISA_REQUIRED",
  CONDITIONAL_WAIVER = "CONDITIONAL_WAIVER",
  ADMISSION_REFUSED = "ADMISSION_REFUSED",
  SPECIAL_TERRITORY = "SPECIAL_TERRITORY",
  UNKNOWN = "UNKNOWN",
}

export enum EntryType {
  SINGLE = "SINGLE",
  MULTIPLE = "MULTIPLE",
  TRANSIT_ONLY = "TRANSIT_ONLY",
  UNSPECIFIED = "UNSPECIFIED",
}

export enum ConditionType {
  REQUIRES_VISA = "REQUIRES_VISA",
  REQUIRES_DOCUMENT = "REQUIRES_DOCUMENT",
  REQUIRES_RESIDENCY = "REQUIRES_RESIDENCY",
  REQUIRES_PURPOSE = "REQUIRES_PURPOSE",
  REQUIRES_ARRIVAL_METHOD = "REQUIRES_ARRIVAL_METHOD",
  REQUIRES_DEPARTURE = "REQUIRES_DEPARTURE",
  AGE_RESTRICTION = "AGE_RESTRICTION",
  INCOME_REQUIREMENT = "INCOME_REQUIREMENT",
  OTHER = "OTHER",
}